feat(pooler): refactor to follow grpcserver pattern integration
