+++ go/cmd/multipooler/plugin_grpcserver.go
@@ -16,4 +16,5 @@ package main
 
 import (
 	_ "github.com/multigres/multigres/go/multipooler/grpcmanagerservice"
+	_ "github.com/multigres/multigres/go/multipooler/grpcpoolerservice"
 )
